# Improve image loading performance