--- a/backend/jvm/src/main/scala/eu/joaocosta/minart/backend/JavaResource.scala
+++ b/backend/jvm/src/main/scala/eu/joaocosta/minart/backend/JavaResource.scala
@@ -1,6 +1,6 @@
 package eu.joaocosta.minart.backend
 
-import java.io.{File, FileInputStream, FileOutputStream, InputStream, OutputStream}
+import java.io.{BufferedInputStream, File, FileInputStream, FileOutputStream, InputStream, OutputStream}
 
 import scala.concurrent._
 import scala.io.Source
@@ -22,7 +22,10 @@ final case class JavaResource(resourcePath: String) extends Resource {
 
   // TODO use Try(Source.fromResource(resourcePath)).getOrElse(Source.fromFile(path)) on scala 2.12+
   def unsafeInputStream(): InputStream =
-    Try(new FileInputStream(path)).orElse(Try(Option(this.getClass().getResourceAsStream("/" + resourcePath)).get)).get
+    Try(new FileInputStream(path))
+      .orElse(Try(Option(this.getClass().getResourceAsStream("/" + resourcePath)).get))
+      .map(is => new BufferedInputStream(is))
+      .get
   def unsafeOutputStream(): OutputStream = new FileOutputStream(path)
 
   def withSourceAsync[A](f: Source => A): Future[A]           = Future(blocking(withSource(f)).get)