--- a/backend/native/src/main/scala/eu/joaocosta/minart/backend/NativeResource.scala
+++ b/backend/native/src/main/scala/eu/joaocosta/minart/backend/NativeResource.scala
@@ -22,9 +22,8 @@ final case class NativeResource(resourcePath: String) extends Resource {
 
   // TODO use Try(Source.fromResource(resourcePath)).getOrElse(Source.fromFile(path)) on scala 2.12+
   def unsafeInputStream(): InputStream =
-    Try(new FileInputStream(path))
+    Try(new BufferedInputStream(new FileInputStream(path)))
       .orElse(Try(Option(this.getClass().getResourceAsStream("/" + resourcePath)).get))
-      .map(is => new BufferedInputStream(is))
       .get
   def unsafeOutputStream(): OutputStream = new FileOutputStream(path)