--- a/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/BmpImageLoader.scala
+++ b/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/BmpImageLoader.scala
@@ -6,8 +6,8 @@ import eu.joaocosta.minart.graphics.RamSurface
 import eu.joaocosta.minart.graphics.image.helpers._
 
 @deprecated("Use eu.joaocosta.minart.graphics.image.bmp.BmpImageFormat instead")
-final class BmpImageLoader[F[_]](val byteReader: ByteReader[F]) extends ImageLoader { self =>
-  private val reader = new bmp.BmpImageReader[F] { val byteReader = self.byteReader }
+final class BmpImageLoader[F[_]](val byteReader: ByteReader[F[Int]]) extends ImageLoader { self =>
+  private val reader = new bmp.BmpImageReader[F[Int]] { val byteReader = self.byteReader }
   def loadImage(is: InputStream): Either[String, RamSurface] = reader.loadImage(is)
 }