--- a/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/PpmImageLoader.scala
+++ b/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/PpmImageLoader.scala
@@ -6,8 +6,8 @@ import eu.joaocosta.minart.graphics.RamSurface
 import eu.joaocosta.minart.graphics.image.helpers._
 
 @deprecated("Use eu.joaocosta.minart.graphics.image.ppm.PpmImageFormat instead")
-final class PpmImageLoader[F[_]](val byteReader: ByteReader[F]) extends ImageLoader { self =>
-  private val reader = new ppm.PpmImageReader[F] { val byteReader = self.byteReader }
+final class PpmImageLoader[F[_]](val byteReader: ByteReader[F[Int]]) extends ImageLoader { self =>
+  private val reader = new ppm.PpmImageReader[F[Int]] { val byteReader = self.byteReader }
   def loadImage(is: InputStream): Either[String, RamSurface] = reader.loadImage(is)
 }