--- a/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/QoiImageLoader.scala
+++ b/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/QoiImageLoader.scala
@@ -6,8 +6,8 @@ import eu.joaocosta.minart.graphics.RamSurface
 import eu.joaocosta.minart.graphics.image.helpers._
 
 @deprecated("Use eu.joaocosta.minart.graphics.image.qoi.QoiImageFormat instead")
-final class QoiImageLoader[F[_]](val byteReader: ByteReader[F]) extends ImageLoader { self =>
-  private val reader = new qoi.QoiImageReader[F] { val byteReader = self.byteReader }
+final class QoiImageLoader[F[_]](val byteReader: ByteReader[F[Int]]) extends ImageLoader { self =>
+  private val reader = new qoi.QoiImageReader[F[Int]] { val byteReader = self.byteReader }
   def loadImage(is: InputStream): Either[String, RamSurface] = reader.loadImage(is)
 }