--- a/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/bmp/BmpImageFormat.scala
+++ b/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/bmp/BmpImageFormat.scala
@@ -11,7 +11,7 @@ final class BmpImageFormat[R, W](val byteReader: ByteReader[R], val byteWriter: 
     with BmpImageWriter[W]
 
 object BmpImageFormat {
-  val defaultFormat = new BmpImageFormat[ByteReader.ModifiableInputStream, Iterator[Array[Byte]]](
+  val defaultFormat = new BmpImageFormat[ByteReader.CustomInputStream, Iterator[Array[Byte]]](
     ByteReader.InputStreamByteReader,
     ByteWriter.IteratorByteWriter
   )