--- a/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/helpers/ByteReader.scala
+++ b/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/helpers/ByteReader.scala
@@ -6,15 +6,15 @@ import scala.collection.compat.immutable.LazyList
 
 /** Helper methods to read binary data from an input stream.
   */
-trait ByteReader[F[_]] {
-  type ParseResult[T]   = Either[String, (F[Int], T)]
-  type ParseState[E, T] = State[F[Int], E, T]
+trait ByteReader[Container] {
+  type ParseResult[T]   = Either[String, (Container, T)]
+  type ParseState[E, T] = State[Container, E, T]
 
   /** Generates a sequence of bytes from a byte stream */
-  def fromInputStream(is: InputStream): F[Int]
+  def fromInputStream(is: InputStream): Container
 
   /** Checks if a byte sequence is empty */
-  def isEmpty[A](seq: F[A]): Boolean
+  def isEmpty(seq: Container): Boolean
 
   /** Adds a sequence of bytes to the head of the byte stream */
   def pushBytes(bytes: Seq[Int]): ParseState[Nothing, Unit]
@@ -50,11 +50,11 @@ trait ByteReader[F[_]] {
 }
 
 object ByteReader {
-  object LazyListByteReader extends ByteReader[LazyList] {
+  object LazyListByteReader extends ByteReader[LazyList[Int]] {
     def fromInputStream(is: InputStream): LazyList[Int] =
       LazyList.continually(is.read()).takeWhile(_ != -1)
 
-    def isEmpty[A](seq: LazyList[A]): Boolean = seq.isEmpty
+    def isEmpty(seq: LazyList[Int]): Boolean = seq.isEmpty
 
     def pushBytes(bytes: Seq[Int]): ParseState[Nothing, Unit] =
       State.modify(s => bytes.to(LazyList) ++ s)
@@ -76,7 +76,7 @@ object ByteReader {
     }
   }
 
-  object IteratorByteReader extends ByteReader[Iterator] {
+  object IteratorByteReader extends ByteReader[Iterator[Int]] {
     // Ported from 2.13 stdlib
     private def nextOption[Int](it: Iterator[Int]): Option[Int] =
       if (it.hasNext) Some(it.next()) else None
@@ -84,7 +84,7 @@ object ByteReader {
     def fromInputStream(is: InputStream): Iterator[Int] =
       Iterator.continually(is.read()).takeWhile(_ != -1)
 
-    def isEmpty[A](seq: Iterator[A]): Boolean = seq.isEmpty
+    def isEmpty(seq: Iterator[Int]): Boolean = seq.isEmpty
 
     def pushBytes(bytes: Seq[Int]): ParseState[Nothing, Unit] =
       State.modify(s => bytes.iterator ++ s)
@@ -121,6 +121,97 @@ object ByteReader {
         bufferedBytes.next()
       }
       bufferedBytes -> buffer.result()
+    }
+  }
+
+  class ModifiableInputStream(inner: InputStream) extends InputStream {
+    val buffer                              = new collection.mutable.Stack[Int]()
+    override def available(): Int           = buffer.size + inner.available()
+    override def close(): Unit              = inner.close()
+    override def mark(readLimit: Int): Unit = ()
+    override def markSupported(): Boolean   = false
+    override def read() = {
+      if (buffer.isEmpty) inner.read()
+      else buffer.pop()
+    }
+    override def read(b: Array[Byte]): Int = {
+      if (buffer.isEmpty || b.isEmpty) inner.read(b)
+      else {
+        val toPop    = math.min(b.size, buffer.size).toInt
+        val toStream = b.size - toPop
+        (0 until toPop).foreach(idx => b(idx) = buffer.pop().toByte)
+        inner.read(b, toPop, toStream) + toPop
+      }
+    }
+    override def reset(): Unit = ()
+    override def skip(n: Long): Long = {
+      if (buffer.isEmpty || n == 0) inner.skip(n)
+      else {
+        val toPop    = math.min(n, buffer.size).toInt
+        val toStream = n - toPop
+        (0 until toPop).foreach(_ => buffer.pop().toByte)
+        inner.skip(toStream) + toPop
+      }
+    }
+
+    // Extra methods
+    def isEmpty(): Boolean =
+      if (buffer.isEmpty) {
+        val next = inner.read()
+        if (next == -1) true
+        else {
+          buffer.push(next)
+          false
+        }
+      } else false
+
+    def pushBytes(bytes: Seq[Int]): this.type = {
+      bytes.reverseIterator.foreach(byte => buffer.push(byte))
+      this
+    }
+  }
+
+  object InputStreamByteReader extends ByteReader[ModifiableInputStream] {
+    def fromInputStream(is: InputStream): ModifiableInputStream = new ModifiableInputStream(is)
+
+    def isEmpty(seq: ModifiableInputStream): Boolean = seq.isEmpty()
+
+    def pushBytes(bytes: Seq[Int]): ParseState[Nothing, Unit] =
+      State.modify(s => s.pushBytes(bytes))
+
+    def skipBytes(n: Int): ParseState[Nothing, Unit] =
+      State.modify { bytes =>
+        bytes.skip(n)
+        bytes
+      }
+
+    val readByte: ParseState[String, Option[Int]] = State { bytes =>
+      val value = bytes.read()
+      bytes -> Option.when(value >= -1)(value)
+    }
+
+    def readBytes(n: Int): ParseState[Nothing, Array[Int]] = State { bytes =>
+      val byteArr    = Array.ofDim[Byte](n)
+      val resultSize = bytes.read(byteArr)
+      bytes -> byteArr.iterator
+        .take(resultSize)
+        .map { byte =>
+          val intValue = byte.toInt
+          if (intValue >= 0) intValue
+          else intValue + 256
+        }
+        .toArray
+    }
+
+    def readWhile(p: Int => Boolean): ParseState[Nothing, List[Int]] = State { bytes =>
+      val buffer = List.newBuilder[Int]
+      var value  = bytes.read()
+      while (value != -1 && p(value)) {
+        buffer += value
+        value = bytes.read()
+      }
+      if (value != -1) bytes.pushBytes(List(value))
+      bytes -> buffer.result()
     }
   }
 }