--- a/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/helpers/ByteReader.scala
+++ b/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/helpers/ByteReader.scala
@@ -195,11 +195,7 @@ object ByteReader {
       val resultSize = bytes.read(byteArr)
       bytes -> byteArr.iterator
         .take(resultSize)
-        .map { byte =>
-          val intValue = byte.toInt
-          if (intValue >= 0) intValue
-          else intValue + 256
-        }
+        .map { byte => java.lang.Byte.toUnsignedInt(byte) }
         .toArray
     }