--- a/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/ppm/PpmImageFormat.scala
+++ b/image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/ppm/PpmImageFormat.scala
@@ -7,12 +7,15 @@ import eu.joaocosta.minart.graphics.image.helpers._
   * Supports reading P2, P3, P5 and P6 PGM/PPM files with a 8 bit color range
   * and stores data as P6 PPM files with a 8 bit color range.
   */
-final class PpmImageFormat[F[_]](val byteReader: ByteReader[F], val byteWriter: ByteWriter[F])
-    extends PpmImageReader[F]
-    with PpmImageWriter[F]
+final class PpmImageFormat[R, W[_]](val byteReader: ByteReader[R], val byteWriter: ByteWriter[W])
+    extends PpmImageReader[R]
+    with PpmImageWriter[W]
 
 object PpmImageFormat {
-  val defaultFormat = new PpmImageFormat[Iterator](ByteReader.IteratorByteReader, ByteWriter.IteratorByteWriter)
+  val defaultFormat = new PpmImageFormat[ByteReader.ModifiableInputStream, Iterator](
+    ByteReader.InputStreamByteReader,
+    ByteWriter.IteratorByteWriter
+  )
 
   val supportedFormats = Set("P2", "P3", "P5", "P6")
 }