… .github/workflows/compile_test.yml
… image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/helpers/ByteWriter.scala
… image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/ppm/PpmImageReader.scala
… image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/ppm/PpmImageWriter.scala
… image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/qoi/QoiColor.scala
… image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/qoi/QoiImageFormat.scala
… image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/qoi/QoiImageReader.scala
… image/shared/src/main/scala/eu/joaocosta/minart/graphics/image/qoi/QoiImageWriter.scala
… image/shared/src/test/resources/lausanne/bmp-24bit.bmp
… image/shared/src/test/resources/lausanne/bmp-32bit.bmp
… image/shared/src/test/resources/lausanne/pgm-p2.pgm
… image/shared/src/test/resources/lausanne/pgm-p5.pgm
… image/shared/src/test/resources/lausanne/ppm-p3.ppm
… image/shared/src/test/resources/lausanne/ppm-p6.ppm
… image/shared/src/test/resources/lausanne/qoi-24bit.qoi
… image/shared/src/test/resources/lausanne/qoi-32bit.qoi
… image/shared/src/test/scala/eu/joaocosta/minart/graphics/image/ImageLoaderSpec.scala
… image/shared/src/test/scala/eu/joaocosta/minart/graphics/image/ImageReaderSpec.scala
… image/shared/src/test/scala/eu/joaocosta/minart/graphics/image/ImageWriterSpec.scala