Improves image loading performance by:
- Accessing the InputStream directly (instead of using an iterator)
- Using a BufferedInputStream
- Reducing memory usage

Most of the performance gains come from the `BufferedInputStream`, but reducing the memory usage is also important to have the code running on scala native.

A 1080p image went from ~6 seconds to ~300ms on the JVM.
The performance numbers were in the same ballpark for all formats, except ASCII based PPMs. Those went from ~30 seconds to ~2 seconds.

This PR also adds a 640x480 image to the test suit to test large images. This is a downscaled photo from [Lausanne by Neil Bates](https://www.pexels.com/photo/snow-light-sea-city-6318902/), distributed as CC0.

(I initially was going to add 1080p images, but that would take ~40MB in resources, which in turn blows the memory usage during scala-native compilation... 9MB seems more than enough)